solve size_t conversion problem
git-svn-id: https://ipvs.informatik.uni-stuttgart.de/SGpp/repos/trunk@266 4eea3252-f0fb-4393-894d-40516dce545b

+++ src/sgpp/basis/modpoly/modified_poly_base.hpp
@@ -129,7 +129,7 @@ private:
 		std::vector<double> y;
 		std::vector<double> intpoly;
 
-		for(int i = 0; i < level - 1; i++)
+		for(size_t i = 0; i < level - 1; i++)
 		{
 			y.push_back(0.0);
 			intpoly.push_back(0.0);
